Added a visual titlescreen and allows player select
Looks for a titlescreen file on the SD card called 'myfile.txt'.
If it exists, it draws the title screen upon load up.
Otherwise, it loads a text title in it's place.

The left and right buttons on the titlescreen will allow you to select
the number of players before you start.